+++ packages/std/src/query/mod.rs
@@ -6,6 +6,7 @@ use crate::Binary;
 use crate::Empty;
 
 mod bank;
+mod distribution;
 mod ibc;
 mod query_response;
 mod staking;
@@ -16,6 +17,8 @@ pub use bank::SupplyResponse;
 pub use bank::{AllBalanceResponse, BalanceResponse, BankQuery};
 #[cfg(feature = "cosmwasm_1_3")]
 pub use bank::{AllDenomMetadataResponse, DenomMetadataResponse};
+#[cfg(feature = "cosmwasm_1_3")]
+pub use distribution::{DelegatorWithdrawAddressResponse, DistributionQuery};
 #[cfg(feature = "stargate")]
 pub use ibc::{ChannelResponse, IbcQuery, ListChannelsResponse, PortIdResponse};
 #[cfg(feature = "staking")]
@@ -35,6 +38,8 @@ pub enum QueryRequest<C> {
     Custom(C),
     #[cfg(feature = "staking")]
     Staking(StakingQuery),
+    #[cfg(feature = "cosmwasm_1_3")]
+    Distribution(DistributionQuery),
     /// A Stargate query is encoded the same way as abci_query, with path and protobuf encoded request data.
     /// The format is defined in [ADR-21](https://github.com/cosmos/cosmos-sdk/blob/master/docs/architecture/adr-021-protobuf-query-encoding.md).
     /// The response is protobuf encoded data directly without a JSON response wrapper.
@@ -106,5 +111,12 @@ impl<C: CustomQuery> From<WasmQuery> for QueryRequest<C> {
 impl<C: CustomQuery> From<IbcQuery> for QueryRequest<C> {
     fn from(msg: IbcQuery) -> Self {
         QueryRequest::Ibc(msg)
+    }
+}
+
+#[cfg(feature = "cosmwasm_1_3")]
+impl<C: CustomQuery> From<DistributionQuery> for QueryRequest<C> {
+    fn from(msg: DistributionQuery) -> Self {
+        QueryRequest::Distribution(msg)
     }
 }